Added max length annotation for model properties.
Added message to controller for max cards for user on Add.

Added async for controller and service methods.

Added call to GetCardsForUserAsync within AddCardAsync and a check for count 50 to limit entries per user. We don't want trolls filling up the DB.

Changed IFlashCardService to accomodate similar changes in implementation.